+++ Sources/SmithyEventStreams/Deserializers/EventContentDeserializer.swift
@@ -72,11 +72,11 @@ struct EventContentDeserializer: ShapeDeserializer {
         throw notImplemented
     }
 
-    func readInteger(_ schema: Schema) throws -> Int {
+    func readInteger(_ schema: Schema) throws -> Int32 {
         throw notImplemented
     }
 
-    func readLong(_ schema: Schema) throws -> Int {
+    func readLong(_ schema: Schema) throws -> Int64 {
         throw notImplemented
     }
